refactor(admin): internalize command runtime
Move the complete Cobra, server, and migration command implementations behind Go's admin/internal boundary. Retain non-executable public compatibility namespaces, route app execution through the internal tree, and add clean external-module positive and negative compilation contracts.
